Update README.md
This project implements a neural network model to classify the gender and approximate age of individuals from images. The model leverages transfer learning with a pre-trained Convolutional Neural Network (CNN) to achieve high accuracy in this task.

The main goals of this project are:

1.Gender classification (Male/Female)
2.Age estimation (categorizing individuals into age ranges)

Dataset
The dataset used in this project consists of images annotated with the gender and age group of individuals.